secops_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-secops-env"
version = "0.1.0"
description = "Cybersecurity Incident Triage RL Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "secops_env.server.app:main"

[tool.setuptools]
packages = ["secops_env", "secops_env.server"]

[tool.setuptools.package-dir]
"secops_env" = "."
"secops_env.server" = "server"
